requirements.txt
certifi==2018.1.18
chardet==3.0.4
dropbox==8.7.1
enum34==1.1.6
futures==3.2.0
httplib2==0.11.3
idna==2.6
isodate==0.6.0
lxml==4.1.1
marcxml-parser==1.2.3
nltk==3.3
numpy==1.14.1
protobuf==3.5.2.post1
psutil==5.4.5
pyDHTMLParser==2.2.2
pymongo==3.6.1
pynlp==0.4.0
PyNLPl==1.2.6
pyparsing==2.2.0
python-docx==0.8.6
rdflib==4.2.2
regex==2018.2.21
remove-hairs==1.0.1
requests==2.18.4
six==1.11.0
stanfordcorenlp==3.9.1.1
urllib3==1.22
virtualenv==15.2.0
xmltodict==0.11.0
